requirements.txt
streamlit==1.31.0
pandas==2.1.4
openpyxl==3.1.2
pytesseract==0.3.10
pillow==10.2.0
easyocr==1.7.1
SpeechRecognition==3.10.1
plotly==5.18.0
gspread==5.12.4
google-auth==2.27.0
pyinstaller==6.3.0
torch==2.1.2
torchvision==0.16.2
opencv-python==4.9.0.80
numpy==1.24.3
scikit-learn==1.4.0
